apis: refactor contract interface creation by adding directly from DB
