--- client/src/data/simpleFiDbTokens.js
@@ -1,39 +1,0 @@
-//Sample shapes for reference
-
-[
-  {
-    token_id: "ed11f92a-7864-4a33-ba7c-acec36e6986b",
-    protocol_id: "8939ce45-8a3e-48f3-beb3-8c09178ead00",
-    name: "Meta",
-    price_api: "mta",
-    address: "0xa3BeD4E1c75D00fa6f4E5E6922DB7261B5E9AcD2",
-  },
-  {
-    token_id: "4c2be9cb-03a4-492f-8e4c-1fd9ced2b41e",
-    protocol_id: null,
-    name: "wETH",
-    price_api: "weth",
-    address: "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
-  },
-  {
-    token_id: "c607a06b-7277-4336-bf2c-aec65fe3f64a",
-    protocol_id: "384acf66-d487-4d39-bd00-6c2ea8518729",
-    name: "MTA-wETH 50/50 receipt",
-    price_api: null,
-    address: "0x0d0d65e7a7db277d3e0f5e1676325e75f3340455",
-  },
-];
-
-Old format:
-{
-  "_id" : ObjectId("5f5940b28153161aff4c55ff"),
-  "name" : "Eth",
-  "apiId" : "ethereum",
-  "transactions" : [
-    {
-      "date" : ISODate("2020-09-09T20:52:46.805Z"),
-      "_id" : ObjectId("5f5940b28153161aff4c5600"),
-      "bought" : true,
-      "price" : 300, "amount" : 2
-    } ],
-    "__v" : 0 }
